Move test file, adjust regex, add test
